# docs/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/users/{username}":{"get":{"tags":["User"],"summary":"Get User","description":"Returns a user from the database.\nSuccess status code: 200\nError status code: 404\n:param db_session: Database session\n:param username: username of the user\n:return: JSON object","operationId":"get_user_users__username__get","parameters":[{"required":true,"schema":{"title":"Username","type":"string"},"name":"username","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/":{"get":{"tags":["User"],"summary":"Get Users","description":"Returns all registered users from the database.\nSuccess status code: 200\nError status code: 400\n:param db_session: Database session\n:return: JSON object","operationId":"get_users_users__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfo"}}}}}},"post":{"tags":["User"],"summary":"Register","description":"Registers the user\n\n:param request: Registration object\n:param response: Response object\n:param db_session: Database session\n:return: JSON object","operationId":"register_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Authenticates the credentials in the request\nReturns a valid JWT token if authentication is successful\nOtherwise, raises an HTTP exception with error code 403\n\n:param credentials: username, password\n:param db: db\n:return: jwt access token","operationId":"login_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Authentication"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/alerts":{"get":{"tags":["Alerts"],"summary":"Get Alerts","description":"Get alerts within the last n days\n:param days: number of days\n:param db: database session\n:param username: username of current user\n:return: list of alerts","operationId":"get_alerts_alerts_get","parameters":[{"required":false,"schema":{"title":"Days","type":"integer","default":1},"name":"days","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Alerts"],"summary":"Create Alert","description":"Create alerts\n:param request: request body\n:param db: database session\n:param username: username of current user\n:return: list of alerts created","operationId":"create_alert_alerts_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertsCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/subscribe":{"post":{"tags":["User"],"summary":"Subscribe","description":"Registers a subscriber in the database.\nSuccess status code: 201\nError status code: 500\n:param username: username of the user\n:param request: request body\n:param db: Database session\n:return: JSON object","operationId":"subscribe_subscribe_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subscriber"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/subscribers":{"get":{"tags":["User"],"summary":"Get Subscribers","description":"Returns all subscribers in the database.\nSuccess status code: 200\nError status code: 500\n:param city: city of the subscriber\n:param pin_code: pin code of the subscriber\n:param email: email of the subscriber\n:param username: username of the user\n:param db: Database session\n:return: JSON object","operationId":"get_subscribers_subscribers_get","parameters":[{"required":false,"schema":{"title":"Email","type":"string"},"name":"email","in":"query"},{"required":false,"schema":{"title":"Pin Code","type":"string"},"name":"pin_code","in":"query"},{"required":false,"schema":{"title":"City","type":"string"},"name":"city","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subscribers"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AlertCreateRequest":{"title":"AlertCreateRequest","required":["title","description","severity"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"severity":{"title":"Severity","type":"string"},"pincodes":{"title":"Pincodes","type":"array","items":{"type":"integer"}},"cities":{"title":"Cities","type":"array","items":{"type":"string"}},"states":{"title":"States","type":"array","items":{"type":"string"}},"countries":{"title":"Countries","type":"array","items":{"type":"string"}},"inform_all":{"title":"Inform All","type":"boolean","default":false}},"description":"Schema for a single alert"},"AlertsCreateRequest":{"title":"AlertsCreateRequest","required":["alerts"],"type":"object","properties":{"alerts":{"title":"Alerts","type":"array","items":{"$ref":"#/components/schemas/AlertCreateRequest"}}},"description":"Schema for alerts creation request"},"Authentication":{"title":"Authentication","required":["username","password"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"}},"description":"Base schema for user authentication"},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"RegistrationRequest":{"title":"RegistrationRequest","required":["username","email","first_name","last_name","password"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"first_name":{"title":"First Name","type":"string"},"last_name":{"title":"Last Name","type":"string"},"password":{"title":"Password","type":"string"}},"description":"Schema for user registration request"},"RegistrationResponse":{"title":"RegistrationResponse","required":["username","email","first_name","last_name"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"first_name":{"title":"First Name","type":"string"},"last_name":{"title":"Last Name","type":"string"}},"description":"Schema for user registration response"},"Subscriber":{"title":"Subscriber","required":["first_name","last_name","email","phone_number","language","pin_code","city","state","country"],"type":"object","properties":{"first_name":{"title":"First Name","type":"string"},"last_name":{"title":"Last Name","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"phone_number":{"title":"Phone Number","type":"string"},"language":{"title":"Language","type":"string"},"pin_code":{"title":"Pin Code","type":"integer"},"city":{"title":"City","type":"string"},"state":{"title":"State","type":"string"},"country":{"title":"Country","type":"string"}},"description":"Schema for a single subscriber"},"Subscribers":{"title":"Subscribers","type":"object","properties":{"subscribers":{"title":"Subscribers","type":"array","items":{"$ref":"#/components/schemas/Subscriber"}}},"description":"Schema for subscribers"},"UserInfo":{"title":"UserInfo","type":"object","properties":{"users":{"title":"Users","type":"array","items":{"$ref":"#/components/schemas/RegistrationResponse"}}},"description":"Schema for users info"},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"login"}}}}}}